MITWPU/DuckieTown Self Driving Car/requirements.txt
opencv-python
numpy
Pillow
Adafruit-SSD1306
Jetson.GPIO
